SQL Import finalization

+++ DBTemplateHander.DatabaseModel.Import/Importer/SQLServerImporterComponents/Models/SQLServerIndexColumnsModel.cs
@@ -10,8 +10,8 @@ namespace DBTemplateHander.DatabaseModel.Import.Importer.SQLServerImporterCompon
         public int index_id { get; set; }
         public int index_column_id { get; set; }
         public int column_id { get; set; }
-        public short key_ordinal { get; set; }
-        public short partition_ordinal { get; set; }
+        public byte key_ordinal { get; set; }
+        public byte partition_ordinal { get; set; }
         public bool is_descending_key { get; set; }
         public bool is_included_column { get; set; }
     }
