+++ DBTemplateHander.DatabaseModel.Import/Importer/SQLServerImporterComponents/Models/SQLServerIndexesModel.cs
@@ -9,14 +9,14 @@ namespace DBTemplateHander.DatabaseModel.Import.Importer.SQLServerImporterCompon
         public int object_id { get; set; }
         public string name { get; set; }
         public int index_id { get; set; }
-        public short type { get; set; }
+        public byte type { get; set; }
         public string type_desc { get; set; }
         public bool? is_unique { get; set; }
         public int? data_space_id { get; set; }
         public bool? ignore_dup_key { get; set; }
         public bool? is_primary_key { get; set; }
         public bool? is_unique_constraint { get; set; }
-        public short fill_factor { get; set; }
+        public byte fill_factor { get; set; }
         public bool? is_padded { get; set; }
         public bool? is_disabled { get; set; }
         public bool? is_hypothetical { get; set; }
